feat(lab): Update cpp01 lab

+++ week01/examples/lab/add.h
@@ -1,3 +1,3 @@
 #pragma once
 
-int add(int n1, int n2);
+long add(long n1, long n2);
